Revised instantiatedGameObject position/rotation/scale to include flipped transform
Revised the math in `StartUp()` to decompose the `transform`'s translation, rotation, and scale all separately and apply those each to the `instantiatedGameObject.transform`.

This fixes an issue where negative scale values applied to the `transform` in `RuleMatches(…)` used to flip tiles weren't being applied to the `instantiatedGameObject`.

This is also more uniform and easier to understand than how I had it before— using `orientationMatrix.MultiplyPoint3x4(…)` for position, `(orientationMatrix * transform)`-derived values for the rotation, and `orientationMatrix.lossyScale` for the scale.  Now the T, R, & S are all `(orientationMatrix * transform)`-derived with just T getting an additional `CellToLocalInterpolated(…)`-based offset.